fix(ui): scope dashboard fullscreen lifecycles
